requirements.txt
Django>=5.2.11
djangorestframework>=3.14.0
requests>=2.31.0
geopy>=2.4.0
python-dotenv>=1.0.0
django-cors-headers>=4.3.0
drf-spectacular>=0.27.0
